requirements.txt
audio-metadata==0.11.1
google-cloud-speech==2.19.0
google-cloud-storage==2.9.0
protobuf==4.23.2